refactor: 게임 결과를 계산하는 기능을 Dealer로 이동
